bamf-matcher: check for matching desktop StartupWMClass and xwindow class
When matching possible desktop applications for window, always check that
the found desktop file has an unset StartupWMClass or a matching one (with the
window's class).

This allow to avoid that applications with same PIDs but with different WMClass
are considered as the same application (fixes chromium applications issues).

(bzr r231.23.5)