Correctly remove the referenced desktop_files from HashTables
BAMF currently uses HashTables which links lists of .desktop files,
when a .desktop file has been (re)moved it should be un-referenced
by these lists, by the way the current implementation didn't do that
correctly. Checking for hashtable values (lists) and not for their 
subvalues (.desktop files).
Of course, this fixes the leakage caused by not removing the desktop
strings and/or the GList structs.

Now it should be fixed.

(bzr r231.30.7)