Matcher: correctly order desktop files by priority.
Desktop files whose name is equal to the desktop_id were prepended
to the desktop file list, but this caused to reverse their priority
(since the ones coming from a less prioritized directory were put
before the ones coming from an high priority directory).

Now the desktop files with a desktop_id matching name are inserted
in a sorted way: they are appended to the other matching desktop files
while they're really prepended to the list of the other desktop files.

(bzr r403.1.14)