BamfMatcher: Correcly match OpenOffice / LibreOffice windows to the right application.
When a window can be mapped as a LibreOffice / OpenOffice application
we look for the right hint for the given window, we set that hint by default
(if valid) and we connect to the window's name-changed event.
So, every time the window name changes, we check again for the window type
and if it has changed (i.e. we moved from writer to start-center), then
we simulate a window open/close event that cause the window to be re-mapped to
the proper application.
This works very well and allows to change the application type on run-time.

Also, every recognized libreoffice window that doesn't match any main
application, is now considered as a "Standard" libreoffice application
window, using the libreoffice-startcenter icon.
This allows to correctly match not only the start center, but also
dialog windows such as the restore window that shows up after a crash.
The slash screen is now ignored, and so doesn't respect the rules above.

Plus some optimizations and code cleanup.

(bzr r423.1.7)