BamfDaemon: rewritten the dbus code to use GDBus.
I'm mostly using code generated by gdbus-codegen, then:

BamfMatcher and BamfControl are now sub-class of the skeleton classes
so, we can avoid code duplication for signals and we can just emit
signals on the object itself or connect to it to handle dbus method calls.

BamfView is instead child of BamfDBusObject and while the views that need
to implement some dbus signals are also implementing the proper interface.
Then each view when initalized add his own interface to the BamfDBusObject.

(bzr r424.1.1)