BamfMatcher: If a Window has changed its class, then we try to rematch it
This is mostly the case of LibreOffice, but if a window has changed its class,
then we should try to re-match, checking if now it can be matched by a new application.
If this is the case, then we remove the window from the previous application (that may
eventually be closed) and we add it to the new one.

Thanks to this we can finally remove a lot of libreoffice-only related code.

(bzr r522.2.3)