BamfFactory: only care about closed views when rematching
It's unsafe to try to rematch an allocated view that is not closed, since it could lead to a
path change. We were not able do this before since we had no way to get if a view
was closed if its remote owner disappeared. With gdbus this is easily possible, instead.

(bzr r424.9.28)